requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
openai
PyMuPDF
tiktoken
chromadb
pdfminer.six